decoders, remote example and wasm experiments

+++ tiff2/src/lib.rs
@@ -8,6 +8,8 @@
 //! * <https://web.archive.org/web/20210108073850/https://www.adobe.io/open/standards/TIFF.html> - The TIFF specification
 //! * <https://download.osgeo.org/libtiff/doc/TIFF6.pdf> - Tiff spec as PDF
 
+#![allow(dead_code, unused_variables)]
+
 /// Generic utility functions that can be used for both decoding and encoding
 pub mod util;
 
@@ -42,32 +44,32 @@ pub enum ChunkType {
 #[derive(Copy, PartialEq, Eq, Debug, Clone, Hash)]
 pub enum ColorType {
     /// Pixel is grayscale
-    Gray(u8),
+    Gray(u16),
 
     /// Pixel contains R, G and B channels
-    RGB(u8),
+    RGB(u16),
 
     /// Pixel is an index into a color palette
-    Palette(u8),
+    Palette(u16),
 
     /// Pixel is grayscale with an alpha channel
-    GrayA(u8),
+    GrayA(u16),
 
     /// Pixel is RGB with an alpha channel
-    RGBA(u8),
+    RGBA(u16),
 
     /// Pixel is CMYK
-    CMYK(u8),
+    CMYK(u16),
 
     /// Pixel is YCbCr
-    YCbCr(u8),
+    YCbCr(u16),
 
     /// Pixel has multiple bands/channels
-    Multiband { bit_depth: u8, num_samples: u16 },
+    Multiband { bit_depth: u16, num_samples: u16 },
 }
 
 impl ColorType {
-    fn bit_depth(&self) -> u8 {
+    fn bit_depth(&self) -> u16 {
         match *self {
             ColorType::Gray(b)
             | ColorType::RGB(b)
